+++ vehicle.py
@@ -1,0 +1,17 @@
+import pygame
+# doi tuong xe luu thong
+class vehicle(pygame.sprite.Sprite):
+    def __init__(self, image, x, y):
+        pygame.sprite.Sprite.__init__(self)
+        # scale img
+        image_scale = 45 / image.get_rect().width
+        new_width = image.get_rect().width * image_scale
+        new_height = image.get_rect().height * image_scale
+        self.image = pygame.transform.scale(image, (new_width, new_height))
+        self.rect = self.image.get_rect()
+        self.rect.center = [x, y]
+# doi tuong xe player
+class PlayerVehicle(vehicle):
+    def __init__(self, x, y):
+        image = pygame.image.load('images/car.png')
+        super().__init__(image, x, y)
